requirements.txt
art>=6.2
python-telegram-bot>=21.3
